dfgc

+++ src/main/java/fr/flowarg/flowupdater/utils/PluginManager.java
@@ -2,7 +2,6 @@ package fr.flowarg.flowupdater.utils;
 
 import fr.antoineok.flowupdater.optifineplugin.Optifine;
 import fr.antoineok.flowupdater.optifineplugin.OptifinePlugin;
-import fr.antoineok.flowupdater.versions.FabricVersion;
 import fr.flowarg.flowio.FileUtils;
 import fr.flowarg.flowlogger.ILogger;
 import fr.flowarg.flowupdater.FlowUpdater;
@@ -10,6 +9,8 @@ import fr.flowarg.flowupdater.curseforgeplugin.CurseForgePlugin;
 import fr.flowarg.flowupdater.curseforgeplugin.CurseMod;
 import fr.flowarg.flowupdater.curseforgeplugin.CurseModPack;
 import fr.flowarg.flowupdater.download.DownloadInfos;
+import fr.flowarg.flowupdater.download.ICurseFeaturesUser;
+import fr.flowarg.flowupdater.download.IProgressCallback;
 import fr.flowarg.flowupdater.download.Step;
 import fr.flowarg.flowupdater.download.json.CurseFileInfos;
 import fr.flowarg.flowupdater.download.json.CurseModPackInfos;
@@ -29,7 +30,7 @@ import static fr.flowarg.flowio.FileUtils.getMD5ofFile;
 
 public class PluginManager
 {
-    private final FlowUpdater updater;
+    private final IProgressCallback progressCallback;
     private final UpdaterOptions options;
     private final ILogger logger;
     private final DownloadInfos downloadInfos;
@@ -39,10 +40,10 @@ public class PluginManager
 
     public PluginManager(FlowUpdater updater)
     {
-        this.updater = updater;
-        this.options = this.updater.getUpdaterOptions();
-        this.logger = this.updater.getLogger();
-        this.downloadInfos = this.updater.getDownloadInfos();
+        this.progressCallback = updater.getCallback();
+        this.options = updater.getUpdaterOptions();
+        this.logger = updater.getLogger();
+        this.downloadInfos = updater.getDownloadInfos();
     }
 
     public void loadPlugins(File dir) throws Exception
@@ -73,10 +74,10 @@ public class PluginManager
         }
     }
 
-    public void loadCurseForgePlugin(File dir, AbstractForgeVersion forgeVersion)
+    public void loadCurseForgePlugin(File dir, ICurseFeaturesUser curseFeaturesUser)
     {
-        final List<Object> allCurseMods = new ArrayList<>(forgeVersion.getCurseMods().size());
-        for (CurseFileInfos infos : forgeVersion.getCurseMods())
+        final List<Object> allCurseMods = new ArrayList<>(curseFeaturesUser.getCurseMods().size());
+        for (CurseFileInfos infos : curseFeaturesUser.getCurseMods())
         {
             if (!this.cursePluginLoaded)
             {
@@ -100,18 +101,21 @@ public class PluginManager
                 final File file = new File(dir, mod.getName());
                 if(!file.exists() || !getMD5ofFile(file).equals(mod.getMd5()) || getFileSizeBytes(file) != mod.getLength())
                 {
-                    file.delete();
-                    this.downloadInfos.getCurseMods().add(mod);
+                    if (!mod.getMd5().contains("-"))
+                    {
+                        file.delete();
+                        this.downloadInfos.getCurseMods().add(mod);
+                    }
                 }
             } catch (Exception e)
             {
                 this.logger.printStackTrace(e);
             }
         }
-        final CurseModPackInfos modPackInfos = forgeVersion.getModPackInfos();
+        final CurseModPackInfos modPackInfos = curseFeaturesUser.getModPackInfos();
         if (modPackInfos != null)
         {
-            this.updater.getCallback().step(Step.MOD_PACK);
+            this.progressCallback.step(Step.MOD_PACK);
             try
             {
                 Class.forName("fr.flowarg.flowupdater.curseforgeplugin.CurseForgePlugin");
@@ -135,89 +139,12 @@ public class PluginManager
                         }
                         if(!flag && (!file.exists() || !getMD5ofFile(file).equals(mod.getMd5()) || getFileSizeBytes(file) != mod.getLength()))
                         {
-                            file.delete();
-                            this.downloadInfos.getCurseMods().add(mod);
-                        }
-                    } catch (NoSuchAlgorithmException | IOException e)
-                    {
-                        this.logger.printStackTrace(e);
-                    }
-                });
-            } catch (ClassNotFoundException e)
-            {
-                this.cursePluginLoaded = false;
-                this.logger.err("Cannot install mod pack from CurseForge: CurseAPI is not loaded. Please, enable the 'enableModsFromCurseForge' updater option !");
-            }
-        }
-
-        forgeVersion.setAllCurseMods(allCurseMods);
-    }
-
-    public void loadCurseForgePlugin(File dir, FabricVersion fabricVersion)
-    {
-        final List<Object> allCurseMods = new ArrayList<>(fabricVersion.getCurseMods().size());
-        for (CurseFileInfos infos : fabricVersion.getCurseMods())
-        {
-            if (!this.cursePluginLoaded)
-            {
-                try
-                {
-                    Class.forName("fr.flowarg.flowupdater.curseforgeplugin.CurseForgePlugin");
-                    this.cursePluginLoaded = true;
-                } catch (ClassNotFoundException e)
-                {
-                    this.cursePluginLoaded = false;
-                    this.logger.err("Cannot install mods from CurseForge: CurseAPI is not loaded. Please, enable the 'enableModsFromCurseForge' updater option !");
-                    break;
-                }
-            }
-
-            try
-            {
-                final CurseForgePlugin curseForgePlugin = CurseForgePlugin.instance;
-                final CurseMod mod = curseForgePlugin.getCurseMod(infos.getProjectID(), infos.getFileID());
-                allCurseMods.add(mod);
-                final File file = new File(dir, mod.getName());
-                if(!file.exists() || !getMD5ofFile(file).equals(mod.getMd5()) || getFileSizeBytes(file) != mod.getLength())
-                {
-                    file.delete();
-                    this.downloadInfos.getCurseMods().add(mod);
-                }
-            } catch (Exception e)
-            {
-                this.logger.printStackTrace(e);
-            }
-        }
-        final CurseModPackInfos modPackInfos = fabricVersion.getModPackInfos();
-        if (modPackInfos != null)
-        {
-            this.updater.getCallback().step(Step.MOD_PACK);
-            try
-            {
-                Class.forName("fr.flowarg.flowupdater.curseforgeplugin.CurseForgePlugin");
-                this.cursePluginLoaded = true;
-                final CurseForgePlugin plugin = CurseForgePlugin.instance;
-                final CurseModPack modPack = plugin.getCurseModPack(modPackInfos.getProjectID(), modPackInfos.getFileID(), modPackInfos.isInstallExtFiles());
-                this.logger.info("Loading mod pack: " + modPack.getName() + " (" + modPack.getVersion() + ") by " + modPack.getAuthor() + '.');
-                modPack.getMods().forEach(mod -> {
-                    allCurseMods.add(mod);
-                    try
-                    {
-                        final File file = new File(dir, mod.getName());
-                        boolean flag = false;
-                        for (String exclude : modPackInfos.getExcluded())
-                        {
-                            if (mod.getName().equalsIgnoreCase(exclude))
+                            if (!mod.getMd5().contains("-"))
                             {
-                                flag = !mod.isRequired();
-                                break;
+                                file.delete();
+                                this.downloadInfos.getCurseMods().add(mod);
                             }
                         }
-                        if(!flag && (!file.exists() || !getMD5ofFile(file).equals(mod.getMd5()) || getFileSizeBytes(file) != mod.getLength()))
-                        {
-                            file.delete();
-                            this.downloadInfos.getCurseMods().add(mod);
-                        }
                     } catch (NoSuchAlgorithmException | IOException e)
                     {
                         this.logger.printStackTrace(e);
@@ -230,7 +157,7 @@ public class PluginManager
             }
         }
 
-        fabricVersion.setAllCurseMods(allCurseMods);
+        curseFeaturesUser.setAllCurseMods(allCurseMods);
     }
 
     public void loadOptifinePlugin(File dir, AbstractForgeVersion forgeVersion)
@@ -272,24 +199,9 @@ public class PluginManager
         this.optifinePluginLoaded = false;
     }
 
-    public FlowUpdater getUpdater()
-    {
-        return this.updater;
-    }
-
-    public UpdaterOptions getOptions()
-    {
-        return this.options;
-    }
-
     public ILogger getLogger()
     {
         return this.logger;
-    }
-
-    public DownloadInfos getDownloadInfos()
-    {
-        return this.downloadInfos;
     }
 
     public boolean isCursePluginLoaded()
